Change reaction_type column type and iconType to tinyint See #2508

+++ wcfsetup/setup/db/install.sql
@@ -1159,9 +1159,9 @@ DROP TABLE IF EXISTS wcf1_reaction_type;
 CREATE TABLE wcf1_reaction_type(
 	reactionTypeID INT(10) NOT NULL AUTO_INCREMENT PRIMARY KEY, 
 	reactionTitle VARCHAR(255), 
-	type SMALLINT(1),
+	type TINYINT(1),
 	showOrder INT(10) NOT NULL DEFAULT 0,
-	iconType SMALLINT(1) DEFAULT 1,
+	iconType TINYINT(1) DEFAULT 1,
 	iconFile MEDIUMTEXT, 
 	iconName VARCHAR(255),
 	iconColor VARCHAR(255),
